refactor: 목업 데이터 이동
-page.tsx -> correctAnswerRate.mock.ts
